requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-dotenv>=0.19.0,<0.20.0
openai>=1.0.0,<2.0.0
streamlit>=1.24.0,<2.0.0
pydantic>=1.8.0,<2.0.0
python-multipart>=0.0.5,<0.0.6
httpx>=0.23.0,<0.24.0
chromadb>=0.3.2,<0.3.3
python-dotenv>=0.19.0,<0.20.0
